Implement 'root view dismiss'
